+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [0.1.4] 2024-8-18
 
+### Added
+
+- Support for DOTNET_ENVIRONMENT environment variable
+
 ### Changed
 
 - Upgraded to .NET 8 and dependencies